fix: 주문 결제 도메인 위치 변경
- 주문 생성부터 시작하도록 변경
