+++ src/main/java/kr/hhplus/be/server/application/order/OrderResult.java
@@ -1,8 +1,6 @@
 package kr.hhplus.be.server.application.order;
 
 import kr.hhplus.be.server.domain.coupon.CouponDto.UserCouponInfo;
-import kr.hhplus.be.server.domain.coupon.UserCoupon;
-import kr.hhplus.be.server.domain.order.Order;
 import kr.hhplus.be.server.domain.order.OrderDto.OrderInfo;
 
 public record OrderResult(
