+++ src/main/java/kr/hhplus/be/server/domain/order/OrderEventPublisher.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.domain.order;
+
+import kr.hhplus.be.server.domain.order.OrderEvent.OrderSuccessEvent;
+
+public interface OrderEventPublisher {
+
+  void success(OrderSuccessEvent event);
+}
