+++ src/main/java/kr/hhplus/be/server/interfaces/order/OrderController.java
@@ -3,10 +3,15 @@ package kr.hhplus.be.server.interfaces.order;
 import jakarta.validation.Valid;
 import kr.hhplus.be.server.application.order.OrderFacade;
 import kr.hhplus.be.server.application.order.OrderResult;
+import kr.hhplus.be.server.application.order.OrderPaymentResult;
 import kr.hhplus.be.server.domain.order.OrderCommand.OrderCreateCommand;
+import kr.hhplus.be.server.domain.payment.PaymentCommand.OrderPaymentCommand;
 import kr.hhplus.be.server.interfaces.CommonResponseWrapper;
+import kr.hhplus.be.server.interfaces.order.OrderRequest.OrderPaymentRequest;
+import kr.hhplus.be.server.interfaces.order.OrderResponse.OrderPaymentResponse;
 import kr.hhplus.be.server.interfaces.order.OrderResponse.OrderSuccessResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -26,6 +31,18 @@ public class OrderController implements OrderControllerSpec {
 
     OrderSuccessResponse response = OrderSuccessResponse.from(orderResult);
 
+    return CommonResponseWrapper.ok(response);
+  }
+
+  @PostMapping("/{orderId}/payments")
+  public CommonResponseWrapper<OrderPaymentResponse> paymentOrder(
+      @PathVariable long orderId,
+      @RequestBody @Valid OrderPaymentRequest request
+  ) {
+    OrderPaymentCommand command = request.toCommand(orderId);
+    OrderPaymentResult orderPaymentResult = orderFacade.payOrder(command);
+
+    OrderPaymentResponse response = OrderPaymentResponse.from(orderPaymentResult);
     return CommonResponseWrapper.ok(response);
   }
 }
