+++ src/main/java/kr/hhplus/be/server/interfaces/order/OrderResponse.java
@@ -1,18 +1,18 @@
 package kr.hhplus.be.server.interfaces.order;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import kr.hhplus.be.server.application.order.OrderPaymentResult;
 import kr.hhplus.be.server.application.order.OrderResult;
 import kr.hhplus.be.server.domain.coupon.CouponDto.UserCouponInfo;
 import kr.hhplus.be.server.domain.coupon.CouponType;
-import kr.hhplus.be.server.domain.coupon.UserCoupon;
-import kr.hhplus.be.server.domain.order.Order;
 import kr.hhplus.be.server.domain.order.OrderDto.OrderInfo;
 import kr.hhplus.be.server.domain.order.OrderDto.OrderItemInfo;
-import kr.hhplus.be.server.domain.order.OrderItem;
 import kr.hhplus.be.server.domain.order.OrderStatus;
+import kr.hhplus.be.server.domain.payment.PaymentStatus;
 
 public record OrderResponse(
 
@@ -144,4 +144,24 @@ public record OrderResponse(
     }
   }
 
+  public record OrderPaymentResponse(
+      @Schema(description = "주문 ID")
+      Long orderId,
+
+      @Schema(description = "유저 ID")
+      Long userId,
+
+      @Schema(description = "주문 총 가격")
+      Long amount
+  ) {
+
+    public static OrderPaymentResponse from(
+        OrderPaymentResult orderPaymentResult) {
+      return new OrderPaymentResponse(
+          orderPaymentResult.orderId(),
+          orderPaymentResult.userId(),
+          orderPaymentResult.finalPrice()
+      );
+    }
+  }
 }
