+++ src/test/java/kr/hhplus/be/server/application/payment/OrderEventListenerTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.verify;
 import kr.hhplus.be.server.domain.payment.PaymentDataClient;
 import kr.hhplus.be.server.domain.payment.PaymentDto.PaymentSuccessPayload;
 import kr.hhplus.be.server.domain.payment.PaymentEvent.PaymentSuccessEvent;
-import kr.hhplus.be.server.interfaces.payment.PaymentEventListener;
+import kr.hhplus.be.server.interfaces.order.OrderEventListener;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -15,13 +15,13 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class PaymentEventListenerTest {
+class OrderEventListenerTest {
 
   @Mock
   private PaymentDataClient paymentDataClient;
 
   @InjectMocks
-  private PaymentEventListener paymentEventListener;
+  private OrderEventListener orderEventListener;
 
   @DisplayName("결제 성공 이벤트를 리슨한다")
   @Test
@@ -30,7 +30,7 @@ class PaymentEventListenerTest {
     PaymentSuccessEvent event = new PaymentSuccessEvent(1L, 1L);
 
     // when
-    paymentEventListener.handlePaymentSuccessEvent(event);
+    orderEventListener.handlePaymentSuccessEvent(event);
 
     // then
     verify(paymentDataClient, atLeastOnce()).publish(PaymentSuccessPayload.from(event));
