+++ src/test/java/kr/hhplus/be/server/e2e/E2EOrderTest.java
@@ -21,9 +21,9 @@ import kr.hhplus.be.server.domain.user.UserTestDataGenerator;
 import kr.hhplus.be.server.interfaces.CommonResponseWrapper;
 import kr.hhplus.be.server.interfaces.order.OrderRequest;
 import kr.hhplus.be.server.interfaces.order.OrderRequest.AmountProductOptionRequest;
+import kr.hhplus.be.server.interfaces.order.OrderRequest.OrderPaymentRequest;
+import kr.hhplus.be.server.interfaces.order.OrderResponse.OrderPaymentResponse;
 import kr.hhplus.be.server.interfaces.order.OrderResponse.OrderSuccessResponse;
-import kr.hhplus.be.server.interfaces.payment.OrderPaymentRequest;
-import kr.hhplus.be.server.interfaces.payment.OrderPaymentResponse;
 import kr.hhplus.be.server.interfaces.point.PointResponse.ChargePointResponse;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -130,7 +130,7 @@ class E2EOrderTest extends IntegrationTestSupport {
     OrderPaymentRequest paymentRequest = new OrderPaymentRequest(userId);
 
     CommonResponseWrapper<OrderPaymentResponse> orderPaymentResponseWrapper = RestAssured.given()
-        .basePath("/api/v1/payments/{orderId}")
+        .basePath("/api/v1/orders/{orderId}/payments")
         .pathParam("orderId", orderResponseWrapper.data().orderId())
         .contentType(ContentType.JSON)
         .body(paymentRequest)
@@ -143,6 +143,6 @@ class E2EOrderTest extends IntegrationTestSupport {
         });
 
     assertThat(orderPaymentResponseWrapper.data().amount()).isGreaterThanOrEqualTo(0L);
-    assertThat(orderPaymentResponseWrapper.data().status()).isEqualTo(PaymentStatus.SUCCESS);
+    assertThat(orderPaymentResponseWrapper.data().orderId()).isNotNull();
   }
 }
